photos-metadata-restore/pyproject.toml
[project]
name = "photos-metadata-restore"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alive-progress>=3.3.0",
    "isort>=6.0.1",
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "pyexiftool>=0.5.1",
    "toml>=0.10.2",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
